+++ deeplearning4j/deeplearning4j-nn/src/main/java/org/deeplearning4j/nn/weights/WeightInitVarScalingNormalFanOut.java
@@ -14,7 +14,7 @@ import org.nd4j.linalg.factory.Nd4j;
 public class WeightInitVarScalingNormalFanOut implements IWeightInit {
 
     @Override
-    public INDArray init(double fanIn, double fanOut, long[] shape, char order, INDArray paramView) {
+    public INDArray init(long fanIn, long fanOut, long[] shape, char order, INDArray paramView) {
         Nd4j.randn(paramView).divi(FastMath.sqrt(fanOut));
         return paramView.reshape(order, shape);
     }
