+++ deeplearning4j/deeplearning4j-nn/src/main/java/org/deeplearning4j/nn/weights/WeightInitVarScalingUniformFanAvg.java
@@ -13,7 +13,7 @@ import org.nd4j.linalg.factory.Nd4j;
 public class WeightInitVarScalingUniformFanAvg implements IWeightInit {
 
     @Override
-    public INDArray init(double fanIn, double fanOut, long[] shape, char order, INDArray paramView) {
+    public INDArray init(long fanIn, long fanOut, long[] shape, char order, INDArray paramView) {
         double scalingFanAvg = 3.0 / Math.sqrt((fanIn + fanOut) / 2);
         Nd4j.rand(paramView, Nd4j.getDistributions().createUniform(-scalingFanAvg, scalingFanAvg));
         return paramView.reshape(order, shape);
